model-backend/requirements.txt
absl-py==0.7.1
astor==0.7.1
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cloudpickle==0.8.1
cycler==0.10.0
dask==1.1.5
decorator==4.4.0
deepexplain @ git+https://github.com/marcoancona/DeepExplain.git@6ac43e729dc8db6ce05ca241dfeb35f1c3eb5b0b
deeplift==0.6.9.0
Flask==1.0.2
Flask-Assets==0.12
Flask-Cors==3.0.7
future==0.17.1
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
idna==2.8
innvestigate @ git+https://github.com/albermax/innvestigate@e76ec2d85a6d59d56b6bbb6dbbf1efee6cea6166
itsdangerous==1.1.0
Jinja2==2.10
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
lime==0.2.0.1
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mock==2.0.0
more-itertools==7.0.0
networkx==2.2
numpy==1.16.2
pandas==0.24.2
pbr==5.1.3
Pillow==5.4.1
pluggy==0.9.0
protobuf==3.7.1
py==1.8.0
pyparsing==2.3.1
pytest==4.3.1
python-dateutil==2.8.0
pytz==2018.9
PyWavelets==1.0.2
PyYAML==5.1
requests==2.21.0
scikit-image==0.14.2
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
toolz==0.9.0
urllib3==1.24.1
webassets==0.12.1
Werkzeug==0.15.1
